大きなファイルサイズに対応できるように修正

+++ amx-t7/t7_amx.c
@@ -105,7 +105,7 @@ static void init_dword8x8(Dword8x8 *mat, int32_t value) {
 }
 
 typedef struct {
-    int cases;
+    int64_t cases;
     int8_t *input_buffer;
     int32_t *result_buffer;
 } TestBuffer;
@@ -118,7 +118,7 @@ static bool read_test_buffer(TestBuffer *test_buffer) {
     }
 
     fread(&test_buffer->cases, sizeof(test_buffer->cases), 1, file);
-    printf("Test cases: %d\n", test_buffer->cases);
+    printf("Test cases: %ld\n", test_buffer->cases);
 
     test_buffer->input_buffer = (int8_t *) malloc(test_buffer->cases * (MAX_SRC_256 * 2) * sizeof(int8_t));
     fread(test_buffer->input_buffer, sizeof(int8_t), test_buffer->cases * (MAX_SRC_256 * 2), file);
